+++ include/vfo_tuner.h
@@ -18,17 +18,17 @@ public:
     virtual void update_display(HT16K33Disp *display);
 
 
-    void frequency_up(float steps);
+    void frequency_up(unsigned long steps);
 
-    void frequency_down(float steps);
+    void frequency_down(unsigned long steps);
 
-    void band_up(float steps);
+    void band_up(unsigned long steps);
 
-    void band_down(float steps);
+    void band_down(unsigned long steps);
 
-    void step_up(float steps);
+    void step_up(unsigned long steps);
 
-    void step_down(float steps);
+    void step_down(unsigned long steps);
 
 private:
 };
